+++ src/components/icon/icon.ts
@@ -71,6 +71,15 @@ export default class IgcIconComponent extends SizableMixin(LitElement) {
   @property({ type: Boolean, reflect: true })
   public mirrored = false;
 
+  /**
+   * The role of icon element.
+   * Defaults to `img`.
+   *
+   * @attr [role="img"]
+   */
+  @property({ reflect: true })
+  public role = 'img';
+
   constructor() {
     super();
     this.size = 'medium';
@@ -78,7 +87,6 @@ export default class IgcIconComponent extends SizableMixin(LitElement) {
 
   public override connectedCallback() {
     super.connectedCallback();
-    this.setAttribute('role', 'img');
     IconsRegistry.instance().subscribe(this.iconLoaded);
   }