+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased
+
+### Added
+- Icon - `role` property which defaults to 'img' [#480](https://github.com/IgniteUI/igniteui-webcomponents/issues/480)
+
 ## [3.3.1] - 2022-08-10
 ### Changed
 - Tree - Removed theme-specified height [#460](https://github.com/IgniteUI/igniteui-webcomponents/pull/460)